% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute missing network data}
\usage{
impute(network, threshold)
}
\arguments{
\item{network}{igraph}

\item{threshold}{dbl}
}
\value{
igraph
}
\description{
This function imputes missing network data using XYZ multiple imputations.
}
